feat: adjust Create Pool button layout next to sort box for small screens (awaiting design confirmation)
